Enter The void
* Biome Added: The Void (used in the new dimension, spawns nothing but
air blocks)
* Dimension Added: The Void, this will be the end game area of this mod
complete with a boss fight, eventually (tm). Currently you spawn in on a
5x5 area made of Voidstone, with one Infused Voidstone in the center as
the spawnpoint. The only way out of this dimension is to use a Void
Pearl. Access to this dimensions is temporarily via the Antenna (right
click it).
* Switched the Void Pearl over to the new Teleporter class
* TODO: add in the Singularity (an item used to open The Void)
* TODO: fix the sea level (maybe?) of the new dimension, for some reason
below Y63 you only see half of the sky